fix(core): fix create select field via createField method
fix #88
